server/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.110.0
uvicorn>=0.29.0
pillow>=10.0.0
numpy>=1.24.0
